Dumping a lot of old code, proper package setup
Between the time I started rewriting the tools, I've learned a lot and
want to make the tools properly, so I'm once again throwing everything
out and starting from scratch.  The old code is sticking around in the
repo for now for easy reference, but will be deleted once it's
replicated.  Also, setup so that it should allow for being a proper
python package.